+++ app/src/main/java/com/ug/air/alrite/Fragments/Patient/ChestIndrawing.java
@@ -302,7 +302,7 @@ public class ChestIndrawing extends Fragment {
 
             String age = sharedPreferences.getString(AGE, "");
             String weight = sharedPreferences.getString(KILO, "");
-            float ag = Float.parseFloat(age);
+            int ag = Integer.parseInt(age);
 
             Instructions instructions = new Instructions();
             messages = instructions.GetPneumoniaInstructions(ag, weight);
