+++ app/src/main/java/com/ug/air/alrite/Fragments/Patient/Nasal.java
@@ -206,7 +206,7 @@ public class Nasal extends Fragment {
         String age = sharedPreferences.getString(AGE, "");
         String s = sharedPreferences.getString(S4, "");
         String weight = sharedPreferences.getString(KILO, "");
-        float ag = Float.parseFloat(age);
+        int ag = Integer.parseInt(age);
 
         Instructions instructions = new Instructions();
         messages = instructions.GetInstructions(ag, weight, s);
