+++ app/src/main/java/com/ug/air/alrite/Fragments/Patient/Oxygen.java
@@ -239,7 +239,7 @@ public class Oxygen extends Fragment {
         String age = sharedPreferences.getString(AGE, "");
         String s = sharedPreferences.getString(S4, "");
         String weight = sharedPreferences.getString(KILO, "");
-        float ag = Float.parseFloat(age);
+        int ag = Integer.parseInt(age);
 
         Instructions instructions = new Instructions();
         messages = instructions.GetInstructions(ag, weight, s);
